Decode array schedule config as composite.
