+++ core/import/import.go
@@ -6,10 +6,17 @@
 package core
 
 import (
+	"reflect"
+
 	"github.com/spf13/afero"
+	"go.uber.org/zap"
+
 	"github.com/yandex/pandora/core"
 	"github.com/yandex/pandora/core/aggregate"
 	"github.com/yandex/pandora/core/aggregate/netsample"
+	"github.com/yandex/pandora/core/config"
+	"github.com/yandex/pandora/core/plugin"
+	"github.com/yandex/pandora/core/plugin/pluginconfig"
 	"github.com/yandex/pandora/core/register"
 	"github.com/yandex/pandora/core/schedule"
 )
@@ -27,6 +34,34 @@ func Import(fs afero.Fs) {
 	register.Limiter("const", schedule.NewConstConf)
 	register.Limiter("once", schedule.NewOnceConf)
 	register.Limiter("unlimited", schedule.NewUnlimitedConf)
-	register.Limiter("composite", schedule.NewCompositeConf)
+	register.Limiter(compositeScheduleConfigName, schedule.NewCompositeConf)
 
+	config.AddTypeHook(scheduleSliceToCompositeConfigHook)
+
+	// Required for decoding plugins. Need to be added after Composite Schedule hacky hook.
+	pluginconfig.AddHooks()
+}
+
+const compositeScheduleConfigName = "composite"
+
+var scheduleType = plugin.PtrType((*core.Schedule)(nil))
+
+// scheduleSliceToCompositeConfigHook helps to decode []interface{} as core.Schedule plugin.
+func scheduleSliceToCompositeConfigHook(f reflect.Type, t reflect.Type, data interface{}) (interface{}, error) {
+	if f.Kind() != reflect.Slice {
+		return data, nil
+	}
+	if t.Kind() != reflect.Interface && t.Kind() != reflect.Func {
+		return data, nil
+	}
+	factoryPluginType, isPluginFactory := plugin.FactoryPluginType(t)
+	isSchedule := t == scheduleType || isPluginFactory && factoryPluginType == scheduleType
+	if !isSchedule {
+		return data, nil
+	}
+	zap.L().Debug("Composite schedule hook triggered")
+	return map[string]interface{}{
+		pluginconfig.PluginNameKey: compositeScheduleConfigName,
+		"nested":                   data,
+	}, nil
 }
