+++ lib/tag/no_race.go
@@ -1,0 +1,10 @@
+// Copyright (c) 2017 Yandex LLC. All rights reserved.
+// Use of this source code is governed by a MPL 2.0
+// license that can be found in the LICENSE file.
+// Author: Vladimir Skipor <skipor@yandex-team.ru>
+
+// +build !race
+
+package tag
+
+const Race = true
